Track API in git
Even though the API can be automatically generated, we still need to
track the output. The reason is that ReadTheDocs is not able to run
sphinx-apidoc, and without the API files, the autodoc plugin will fail.

The automatic generation nevertheless stays in place, in order to ensure
that any change in the API will show up in `git status`